+++ NEWS.md
@@ -1,5 +1,11 @@
 # cargo-mutants changelog
 
+## 23.11.1
+
+- New `--in-diff FILE` option tests only mutants that are in the diff from the
+  given file. This is useful to avoid testing mutants from code that has not changed,
+  either locally or in CI.
+
 ## 23.11.0
 
 - Changed: `cargo mutants` now tries to match the behavior of `cargo test` when run within a workspace. If run in a package directory, it tests only that package. If run in a workspace that is not a package (a "virtual workspace"), it tests the configured default packages, or otherwise all packages. This can all be overridden with the `--package` or `--workspace` options.